move find_git_root to utils/_locate, tweak formatting
